+++ src/babel/__tests__/fixtures/imports-globals.js
@@ -3,7 +3,7 @@ const event = createEvent()
 const store = createStore(0)
 const effect = createEffect()
 const domainEvent = domain.createEvent()
-const scope = fork(domain)
+const scope = fork()
 const node = createNode()
 
 const a = attach({effect})